# public/swagger.yml
openapi: 3.0.3
info:
  title: Users with todo lists
  version: 1.0.0
servers: 
  - url: http://localhost:3000/
paths:
  /api/auth/registration:
    post:
      summary: Register user
      tags:
        - Auth
      operationId: registration
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                username:
                  type: string
                  example: sergo
                password:
                  type: string
                  example: sergo
                role:
                  type: string
                  enum:
                    - USER
                    - ADMIN
                  description: Access rights
                email:
                  type: string
                  example: sergo@gmail.com
                firstname:
                  type: string
                  example: Serhii
                lastname:
                  type: string
                  example: Pedos
      responses:
        '201':
          description: Successful response with a registered user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        '400':
          description: "User already exists"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error400"
        '404':
          description: "User not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error404"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    User:
      type: object
      required:
        - username
        - password
        - role
        - email
        - firstname
        - lastname
      properties:
        _id:
          type: string
          example: 6414218570da02fed5d5eb6b
        username:
          type: string
          example: sergo
        password:
          type: string
          example: sergo
        role:
          type: string
          enum:
            - USER
            - ADMIN
          description: Access rights
        email:
          type: string
          example: sergo@gmail.com
        firstname:
          type: string
          example: Serhii
        lastname:
          type: string
          example: Pedos
        createdAt:
          type: string
          example: 2023-03-17T08:15:16.068+00:00
          format: 2023-03-17T08:15:16.068+00:00
...
    Error400:
      type: object
      required:
        - code
        - name
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: Bad Request
...
    Error404:
      type: object
      required:
        - code
        - name
      properties:
        code:
          type: integer
          example: 404
        message:
          type: string
          example: Not found
security:
  - bearerAuth: []